Fixes retrieving Error subclasses from user info
std::any requires we cast back to the exact type that was stored, which means we can’t store an Error subclass and then retrieve it as Error.

We work around this by storing all underlying errors as std::shared_ptr<Error>, which suppports polymorphism but is the same static type as far as std::any is concerned.